requirements.txt
numpy
faiss-cpu
sentence-transformers
streamlit